cybersecurity-social-media-post/requirements.txt
tweepy==4.10.0
requests==2.28.1
python-dotenv==0.21.0  # For managing environment variables

